+++ entity/constants.py
@@ -1,4 +1,0 @@
-
-class MembershipType:
-    UNION = 'UNION'
-    INTERSECTION = 'INTERSECTION'